added validation and change in put mapping method

+++ src/main/java/my/mood/ExpenceTracker/Expence_Tracker/Expence/ExpenceController.java
@@ -8,8 +8,11 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+
+import jakarta.validation.Valid;
 
 @RestController
 public class ExpenceController {
@@ -40,13 +43,43 @@ public class ExpenceController {
     @DeleteMapping("/expences/{id}")
     public void deleteSpecificExpence(@PathVariable int id) {
          // expenceService.deleteFromId(id);
-    	expenceRepository.deleteById(id);
+    	Optional<Expence_Entity> existingExpence = expenceRepository.findById(id);
+    	
+    	if(existingExpence.isPresent()) {
+    		expenceRepository.deleteById(id);
+    	}
+    	
+    	else {
+    		throw new RuntimeException("Expence not found with ID = " + id); 
+    	}
     }
 
     @PostMapping("/expences")
-    public Expence_Entity addExpence(@RequestBody Expence_Entity entity) {
+    public Expence_Entity addExpence(@Valid @RequestBody Expence_Entity entity) {
         // return expenceService.addExpence(entity);
     	return expenceRepository.save(entity);
+    }
+    
+    @PutMapping("/expences/{id}")
+    public Expence_Entity updateExpence(@PathVariable int id,  @Valid @RequestBody Expence_Entity entity) {
+    	Optional<Expence_Entity> existingExpence = expenceRepository.findById(id);
+    	
+    	if(existingExpence.isPresent()) {
+    		Expence_Entity expence = existingExpence.get();
+    		
+    		expence.setId(id);
+    		expence.setTitle(entity.getTitle());
+    		expence.setAmount(entity.getAmount());
+    		expence.setCategory(entity.getCategory());
+    		expence.setDate(entity.getDate());
+    	
+        	return expenceRepository.save(expence);
+    	}
+    	
+    	else {
+    		throw new RuntimeException("Expence not found with ID = " + id); 
+    	}
+    	
     }
     
 }
